+++ test/pytests/test_smart_completion_public_schema_only.py
@@ -773,6 +773,15 @@ def dummy_list_path(dir_name):
         ('source --special /', [('/dir1', -1), ('/file1.sql', -1), ('/file2.sql', -1)]),
         ('source --show /', [('/dir1', -1), ('/file1.sql', -1), ('/file2.sql', -1)]),
         (
+            'source file.sql ',
+            [('--special', 0), ('--show', 0), ('--page', 0), ('--throttle', 0), ('--help', 0)],
+        ),
+        (
+            'source file.sql --show ',
+            [('--special', 0), ('--page', 0), ('--throttle', 0), ('--help', 0)],
+        ),
+        ('source -- ', [('/', 0), ('~', 0), ('.', 0), ('..', 0)]),
+        (
             "source /dir1/",
             [("/dir1/subdir1", -6), ("/dir1/subfile1.sql", -6), ("/dir1/subfile2.sql", -6)],
         ),